Display turn order as text
The game creates 5 text boxes to display the turn order.

The game manager runs a predictive initiative tick which affects the predictive initiative values in the character stats class, rather than the actual initiative values.

It then creates a list of the 5 next characters and saves it to the database which triggers an event to inform the UI.

The UI then displays this information by putting the name of each character into the text boxes.

The game manager still needs to actually update the turn order when a new character is taken.