Turn order predicted
Each turn, the game manager will calculate the next 5 characters assuming they take a full turn, rather than skipping.

(It actually calculates 6 characters but removes the first as this will be the current character, leaving 5 more to be displayed).

At the start of the turn, the predicted initiative values for each character are reset to their current values. the previous "currentTurnOrder" is emptied and the whole calculation happens again.