Popups
Big overhaul to how popups work.

Any widget that can have a popup inherits from popupable.

when a widget is hovered on for a set time, it sends notifies the database and triggeres an event.

The popup canvas listens to that event and displays the popup with the data from the database.

The popup canvas checks if either the widget or the popup are being hovered on. When they're no long hovered, it removes them.

When the popup is removed from the database, it calls a function on the widget that called for it which allows the widget to call for another popup next time it is hovered on.